fixed warning... thanks to Georg for identifying...
git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@21994 060e22d8-9d1c-0410-868a-e3a950d63d95

+++ src/drt_scatra/scatra_resulttest.cpp
@@ -267,7 +267,7 @@ double SCATRA::ScaTraResultTest::ResultSpecial(
 
     // extract relevant result from map
     std::map<const int,std::vector<double> >::const_iterator iterator = map->find(domain);
-    if(iterator == map->end() or species < 0 or species >= iterator->second.size())
+    if(iterator == map->end() or species < 0 or species >= (int) iterator->second.size())
       dserror("Couldn't extract total or mean value of transported scalar with ID %d inside domain with ID %d from time integrator!",species,domain);
     result = iterator->second[species];
   }
